requirements.txt
evadb==0.3.7
gpt4all==1.0.12
openai==1.3.5
python-dotenv==1.0.0
# mysqlclient==2.2.0
# psycopg2==2.9.9
# psycopg2-binary==2.9.9